Refactor filter/refresh buttons
Filter button is now referred to as Query since that's the name of the variable
in the collection. Also switched this to a modal to keep it consistent with
other forms and not being able to know exactly which direction the previous
Popover should go.

Also made use of the AdminLTE "App Buttons" since the badge gives an extra place
to put information. The "App Button" also is consistent with the "Edit" and
"Delete".